requirements.txt
langchain
langchain-core
langchain-community
langgraph

langchain-google-genai

pydantic

python-dotenv